Target commonjs for transpiled output
There are numerous issues with targeting ES Modules:

Importing TypeScript packages transpiled to commonjs does not behave the
same, most troubeling is that the default export when imported becomes
the module.exports of the transpiled output instead of the default
export declared in the source file.

Jest support for ES Modules is experimental and not usable.  Module
loading does not correctly resolve packages with the "exports" field in
package.json. And loading many modules in parallel with import breaks
the whole instrumentation. The API used by Jest to support ES Modules is
also behind the --experimental-vm-modules flag and not stable.  See
issue 9430 in the jest repositiory for details on the support.

Targeting ES Modules is for these reasons not not practical.